break out `stringToAutocompleteObject` and `createInitialValue` into a utils file, add specs
